feat(perfectionist): update options `partitionByComment` for all related rules
