+++ src/eslint/configs.ts
@@ -1,5 +1,5 @@
 /**
- * Configs
+ * @file Configs
  */
 
 export { configs as configsTypescript } from 'typescript-eslint'
