+++ src/types/eslint.ts
@@ -4,11 +4,6 @@ import type { RuleOptions } from './typegen'
 import type { Pretty } from './utils'
 
 /**
- * ESLint rules
- */
-export type ESLintRulesRecord = Linter.RulesRecord
-
-/**
  * ESLint config
  */
 export type ESLintConfig<Rules extends ESLintRulesRecord = ESLintRulesRecord> = Linter.Config<Rules>
@@ -19,14 +14,14 @@ export type ESLintConfig<Rules extends ESLintRulesRecord = ESLintRulesRecord> = 
 export type ESLintParser = Linter.Parser
 
 /**
- * ESLint processor
- */
-export type ESLintProcessor = Linter.Processor
-
-/**
  * ESLint parserOptions
  */
 export type ESLintParserOptions = Linter.ParserOptions
+
+/**
+ * ESLint processor
+ */
+export type ESLintProcessor = Linter.Processor
 
 /**
  * ESLint rule severity
@@ -34,6 +29,11 @@ export type ESLintParserOptions = Linter.ParserOptions
  * for config options use, don't need `off`
  */
 export type ESLintRuleSeverity = 'error' | 'warn'
+
+/**
+ * ESLint rules
+ */
+export type ESLintRulesRecord = Linter.RulesRecord
 
 /**
  * TypeScript ESLint parserOptions
@@ -45,7 +45,6 @@ export type TSESLintParserOptions = Pretty<
 /**
  * Typed flat config item
  */
-
 export type TypedConfigItem = Omit<Linter.Config<ESLintRulesRecord & RuleOptions>, 'plugins'> & {
   /**
    * Most plugin are not properly typed
