+++ src/types/shared.ts
@@ -1,0 +1,11 @@
+/**
+ * @see {@link https://perfectionist.dev/rules/sort-imports#partitionbycomment}
+ */
+export type PerfectionistPartitionByComment =
+  | boolean
+  | string
+  | string[]
+  | {
+      block?: boolean | string | string[]
+      line?: boolean | string | string[]
+    }
